# Laboratory1/1016106015/Swagger Documentation/Swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
        "/notifications/send": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Env\u00eda una notificaci\u00f3n usando la cadena de responsabilidad",
                "operationId": "post_send_notification",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Notification"
                        }
                    }
                ],
                "tags": [
                    "notifications"
                ]
            }
        },
        "/users": {
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Lista todos los usuarios",
                "operationId": "get_user_list",
                "tags": [
                    "users"
                ]
            },
            "post": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "summary": "Registra un usuario con canales",
                "operationId": "post_user_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        }
    },
    "info": {
        "title": "Multichannel Notification API",
        "version": "1.0",
        "description": "Env\u00eda notificaciones por email/SMS con fallback"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "tags": [
        {
            "name": "users",
            "description": "Gesti\u00f3n de usuarios"
        },
        {
            "name": "notifications",
            "description": "Env\u00edo de notificaciones"
        }
    ],
    "definitions": {
        "User": {
            "required": [
                "available_channels",
                "name",
                "preferred_channel"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "preferred_channel": {
                    "type": "string",
                    "example": "email",
                    "enum": [
                        "email",
                        "sms"
                    ]
                },
                "available_channels": {
                    "type": "array",
                    "description": "Lista de canales ['email','sms']",
                    "items": {
                        "type": "string"
                    }
                }
            },
            "type": "object"
        },
        "Notification": {
            "required": [
                "message",
                "priority",
                "user_name"
            ],
            "properties": {
                "user_name": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "priority": {
                    "type": "string",
                    "example": "low",
                    "enum": [
                        "low",
                        "medium",
                        "high"
                    ]
                }
            },
            "type": "object"
        }
    },
    "responses": {
        "ParseError": {
            "description": "When a mask can't be parsed"
        },
        "MaskError": {
            "description": "When any error occurs on mask"
        }
    }
}